Changes after Tommie's first testing

+++ HOTP/Controllers/Plan90Controller.cs
@@ -7,6 +7,14 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using HOTP.Models;
+using System.IO;
+//using RazorPDF;
+using iTextSharp;
+using iTextSharp.text;
+using iTextSharp.text.pdf;
+using System.Xml;
+using System.Text;
+
 
 namespace HOTP.Controllers
 {
@@ -88,12 +96,29 @@ namespace HOTP.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            tblHOTP_Plan90 tblHOTP_Plan90 = db.tblHOTP_Plan90.Find(id);
-            if (tblHOTP_Plan90 == null)
-            {
-                return HttpNotFound();
-            }
-            return View(tblHOTP_Plan90);
+            PopulatePlan(id);
+            tblHOTP_EmployeeGoals employeeGoal = db.tblHOTP_EmployeeGoals.Find(id);
+            //var goal = db.tblHOTP_Goals.Find(employeeGoal.GoalID);
+            tblHOTP_Employees employee = db.tblHOTP_Employees.Find(employeeGoal.EmployeeID);
+
+            var plans = from p in db.tblHOTP_Plan90
+                        where p.EmployeeGoalID == id
+                        join eg in db.tblHOTP_EmployeeGoals on p.EmployeeGoalID equals eg.EmployeeGoalID into empGoal
+                        from subEmployeeGoal in empGoal.DefaultIfEmpty()
+                        join g in db.tblHOTP_Goals on employeeGoal.GoalID equals g.GoalID into goal
+                        from subGoal in goal.DefaultIfEmpty()
+                        select new Plan90ViewModel
+                        {
+                            EmployeeName = employee.FirstName + " " + employee.LastName,
+                            EmployeeGoalID = employeeGoal.EmployeeGoalID,
+                            Weight = employeeGoal.Weight,
+                            ItemScore = employeeGoal.ItemScore,
+                            Goal = subGoal,
+                            Plan = p,
+                            ActionSteps = db.tblHOTP_ActionSteps.Where(a => a.PlanID == p.PlanID).ToList()
+                        };
+            ViewBag.PlanStatus = PopulateCodesDDL("PlanStatus");
+            return View(plans.ToList());
         }
 
 
@@ -281,55 +306,39 @@ namespace HOTP.Controllers
             TempData["YearEnding"] = YearEnding;
             ViewBag.YearEnding = PopulateCodesDDL("YearEnding", YearEnding);
             ViewBag.PlanStatus = PopulateCodesDDL("PlanStatus");
-
             return View(GetAllGoals(SelectedEmp, YearEnding));
-
-            //List<GoalFull> AllGoals = new List<GoalFull>();
-
-            //if (SelectedEmp != null)
-            //{
-            //    TempData["SelectedEmp"] = SelectedEmp;
-
-            //    var emp = (from e in db.tblHOTP_Employees
-            //               where e.EmployeeID == SelectedEmp
-            //               select e).First();
-
-            //    var empGoals = from eg in db.tblHOTP_EmployeeGoals
-            //                   join g in db.tblHOTP_Goals on eg.GoalID equals g.GoalID into goal
-            //                   from subGoal in goal.DefaultIfEmpty()
-            //                   where eg.EmployeeID == SelectedEmp && subGoal.YearEnding == YearEnding
-            //                   select eg;
-            //    foreach (tblHOTP_EmployeeGoals eg in empGoals.ToList())
-            //    {
-            //        PopulatePlan(eg.EmployeeGoalID);
-            //    }
-
-            //    foreach (tblHOTP_EmployeeGoals eg in empGoals)
-            //    {
-            //        List<Plan90Full> gPlans = new List<Plan90Full>();
-            //        var goalPlans = from plan in db.tblHOTP_Plan90 where plan.EmployeeGoalID == eg.EmployeeGoalID select plan;
-            //        foreach (tblHOTP_Plan90 plan in goalPlans)
-            //        {
-            //            gPlans.Add(new Plan90Full()
-            //            {
-            //                Plan = plan,
-            //                ActionSteps = db.tblHOTP_ActionSteps.Where(a => a.PlanID == plan.PlanID).ToList()
-            //            });
-            //        }
-            //        AllGoals.Add(new GoalFull()
-            //        {
-            //            EmployeeName = emp.FirstName + " " + emp.LastName,
-            //            Goal = db.tblHOTP_Goals.Find(eg.GoalID),
-            //            EmployeeGoal = eg,
-            //            Plans = gPlans
-            //        });
-            //    }
-            //}
-            //return View(AllGoals.ToList());
         }
+
 
         // GET: Plan90ViewModel/Report/5
         public ActionResult Report(int? SelectedEmp, string YearEnding)
+        {
+            if (SelectedEmp == null && TempData["SelectedEmp"] != null)
+            {
+                SelectedEmp = Convert.ToInt16(TempData["SelectedEmp"]);
+            }            
+            var emps =
+              db.tblHOTP_Employees
+                .Where(s => s.Evaluations)
+                .OrderBy(s => s.LastName)
+                .ToList()
+                .Select(s => new
+                {
+                    EmployeeId = s.EmployeeID,
+                    FullName = string.Format("{0}, {1}", s.LastName, s.FirstName)
+                });
+            ViewBag.SelectedEmp = new SelectList(emps, "EmployeeID", "FullName", SelectedEmp);
+            if (YearEnding == null && TempData["YearEnding"] != null)
+            {
+                YearEnding = TempData["YearEnding"].ToString();
+            }
+            TempData["YearEnding"] = YearEnding;
+            TempData["SelectedEmp"] = SelectedEmp;
+            return View(GetAllGoals(SelectedEmp, YearEnding));
+        }
+
+        // GET: Plan90ViewModel/Report/5
+        public ActionResult Report2(int? SelectedEmp, string YearEnding)
         {
             if (SelectedEmp == null && TempData["SelectedEmp"] != null)
             {
@@ -352,53 +361,116 @@ namespace HOTP.Controllers
                 YearEnding = TempData["YearEnding"].ToString();
             }
             TempData["YearEnding"] = YearEnding;
-            ViewBag.YearEnding = PopulateCodesDDL("YearEnding", YearEnding);
-            ViewBag.PlanStatus = PopulateCodesDDL("PlanStatus");
+            //ViewBag.YearEnding = PopulateCodesDDL("YearEnding", YearEnding);
+            //ViewBag.PlanStatus = PopulateCodesDDL("PlanStatus");
+            //return View(GetAllGoals(SelectedEmp, YearEnding));
+            return new RazorPDF.PdfResult(GetAllGoals(SelectedEmp, YearEnding), "Report");
+        }
 
+        //protected string RenderActionResultToString(ActionResult result)
+        //{
+        //    // Create memory writer.
+        //    var sb = new StringBuilder();
+        //    var memWriter = new StringWriter(sb);
+
+        //    // Create fake http context to render the view.
+        //    var fakeResponse = new HttpResponse(memWriter);
+        //    var fakeContext = new HttpContext(System.Web.HttpContext.Current.Request,
+        //        fakeResponse);
+        //    var fakeControllerContext = new ControllerContext(
+        //        new HttpContextWrapper(fakeContext),
+        //        this.ControllerContext.RouteData,
+        //        this.ControllerContext.Controller);
+        //    var oldContext = System.Web.HttpContext.Current;
+        //    System.Web.HttpContext.Current = fakeContext;
+
+        //    // Render the view.
+        //    result.ExecuteResult(fakeControllerContext);
+
+        //    // Restore old context.
+        //    System.Web.HttpContext.Current = oldContext;
+
+        //    // Flush memory and return output.
+        //    memWriter.Flush();
+        //    return sb.ToString();
+        //}
+
+        //protected ActionResult ViewPdf(object model)
+        //{
+        //    // Create the iTextSharp document.
+        //    iTextSharp.text.Document doc = new Document();
+        //    // Set the document to write to memory.
+        //    MemoryStream memStream = new MemoryStream();
+        //    PdfWriter writer = PdfWriter.GetInstance(doc, memStream);
+        //    writer.CloseStream = false;
+        //    doc.Open();
+
+        //    // Render the view xml to a string, then parse that string into an XML dom.
+        //    string xmltext = this.RenderActionResultToString(this.View(model));
+        //    XmlDocument xmldoc = new XmlDocument();
+        //    xmldoc.InnerXml = xmltext.Trim();
+
+        //    // Parse the XML into the iTextSharp document.
+        //    iTextSharp.tool.xml.XMLWorker textHandler = new iTextSharp.tool.xml.XMLWorker(doc);
+
+        //     textHandler = new ITextHandler(doc);
+        //    textHandler.Parse(xmldoc);
+
+        //    // Close and get the resulted binary data.
+        //    doc.Close();
+        //    byte[] buf = new byte[memStream.Position];
+        //    memStream.Position = 0;
+        //    memStream.Read(buf, 0, buf.Length);
+
+        //    // Send the binary data to the browser.
+        //    return new BinaryContentResult(buf, "application/pdf");
+        //}
+
+
+
+        public ActionResult ReportQuarter()
+        {
+            int SelectedEmp = 1;
+            string YearEnding = "2005";
             return View(GetAllGoals(SelectedEmp, YearEnding));
+        }
 
-            //List<GoalFull> AllGoals = new List<GoalFull>();
+//        public ActionResult DownloadViewPDF()
+//{
+//var model = new GeneratePDFModel();
+////Code to get content
+//return new Rotativa.ViewAsPdf("GeneratePDF", model){FileName = "TestViewAsPdf.pdf"}
+//}
 
-            //if (SelectedEmp != null)
-            //{
-            //    TempData["SelectedEmp"] = SelectedEmp;
+        //public ActionResult DownloadActionAsPDF()
+        //{
+        //    var model = new GeneratePDFModel();
+        //    //Code to get content
+        //    return new Rotativa.ActionAsPdf("GeneratePDF", model) { FileName = "TestActionAsPdf.pdf" };
+        //}
 
-            //    var emp = (from e in db.tblHOTP_Employees
-            //               where e.EmployeeID == SelectedEmp
-            //               select e).First();
+        //public ActionResult GeneratePDF()
+        //{
+        //    var model = new GeneratePDFModel();
+        //    //get content
+        //    return View(model);
+        //}
 
-            //    var empGoals = from eg in db.tblHOTP_EmployeeGoals
-            //                   join g in db.tblHOTP_Goals on eg.GoalID equals g.GoalID into goal
-            //                   from subGoal in goal.DefaultIfEmpty()
-            //                   where eg.EmployeeID == SelectedEmp && subGoal.YearEnding == YearEnding
-            //                   select eg;
-            //    foreach (tblHOTP_EmployeeGoals eg in empGoals.ToList())
-            //    {
-            //        PopulatePlan(eg.EmployeeGoalID);
-            //    }
+//        public ActionResult DownloadPartialViewPDF()
+//        {
+//            var model = new GeneratePDFModel();
+//            //Code to get content
+//            return new Rotativa.PartialViewAsPdf("_PartialViewTest", model) { FileName = "TestPartialViewAsPdf.pdf" };
+//        }
 
-            //    foreach (tblHOTP_EmployeeGoals eg in empGoals)
-            //    {
-            //        List<Plan90Full> gPlans = new List<Plan90Full>();
-            //        var goalPlans = from plan in db.tblHOTP_Plan90 where plan.EmployeeGoalID == eg.EmployeeGoalID select plan;
-            //        foreach (tblHOTP_Plan90 plan in goalPlans)
-            //        {
-            //            gPlans.Add(new Plan90Full()
-            //            {
-            //                Plan = plan,
-            //                ActionSteps = db.tblHOTP_ActionSteps.Where(a => a.PlanID == plan.PlanID).ToList()
-            //            });
-            //        }
-            //        AllGoals.Add(new GoalFull()
-            //        {
-            //            EmployeeName = emp.FirstName + " " + emp.LastName,
-            //            Goal = db.tblHOTP_Goals.Find(eg.GoalID),
-            //            EmployeeGoal = eg,
-            //            Plans = gPlans
-            //        });
-            //    }
-            //}
-            //return View(AllGoals.ToList());
+        public ActionResult UrlAsPDF()
+        {
+            return new Rotativa.UrlAsPdf("/Report") { FileName = "UrlTest.pdf" };
+        }
+
+        public ActionResult GeneratePDF()
+        {
+            return new Rotativa.ViewAsPdf("ReportQuarter", GetAllGoals(1, "2015"));
         }
 
 
@@ -417,33 +489,36 @@ namespace HOTP.Controllers
                     {
                         tblHOTP_Plan90 existingPlan = db.tblHOTP_Plan90.Find(plan90.Plan.PlanID);
                         existingPlan.Goal = plan90.Plan.Goal;
-                        foreach (tblHOTP_ActionSteps actionStep in plan90.ActionSteps)
+                        if (plan90.ActionSteps != null)
                         {
-                            if (actionStep.ActionID != 99999 && actionStep.Status == "Remove")
+                            foreach (tblHOTP_ActionSteps actionStep in plan90.ActionSteps)
                             {
-                                tblHOTP_ActionSteps existingAction = db.tblHOTP_ActionSteps.Find(actionStep.ActionID);
-                                db.tblHOTP_ActionSteps.Remove(existingAction);
-                            }
-                            else
-                                if (actionStep.ActionID == 99999 && actionStep.Status != "Remove")
+                                if (actionStep.ActionID != 99999 && actionStep.Status == "Remove")
                                 {
-                                    tblHOTP_ActionSteps newActionStep = new tblHOTP_ActionSteps();
-                                    newActionStep.PlanID = plan90.Plan.PlanID;
-                                    newActionStep.ActionStep = actionStep.ActionStep;
-                                    newActionStep.Result = actionStep.Result;
-                                    newActionStep.Status = actionStep.Status;
-                                    db.tblHOTP_ActionSteps.Add(newActionStep);
+                                    tblHOTP_ActionSteps existingAction = db.tblHOTP_ActionSteps.Find(actionStep.ActionID);
+                                    db.tblHOTP_ActionSteps.Remove(existingAction);
                                 }
                                 else
-                                {
-                                    if (actionStep.ActionID != 99999)
+                                    if (actionStep.ActionID == 99999 && actionStep.Status != "Remove")
                                     {
-                                        tblHOTP_ActionSteps existingAction = db.tblHOTP_ActionSteps.Find(actionStep.ActionID);
-                                        existingAction.Result = actionStep.Result;
-                                        existingAction.ActionStep = actionStep.ActionStep;
-                                        existingAction.Status = actionStep.Status;
+                                        tblHOTP_ActionSteps newActionStep = new tblHOTP_ActionSteps();
+                                        newActionStep.PlanID = plan90.Plan.PlanID;
+                                        newActionStep.ActionStep = actionStep.ActionStep;
+                                        newActionStep.Result = actionStep.Result;
+                                        newActionStep.Status = actionStep.Status;
+                                        db.tblHOTP_ActionSteps.Add(newActionStep);
                                     }
-                                }
+                                    else
+                                    {
+                                        if (actionStep.ActionID != 99999)
+                                        {
+                                            tblHOTP_ActionSteps existingAction = db.tblHOTP_ActionSteps.Find(actionStep.ActionID);
+                                            existingAction.Result = actionStep.Result;
+                                            existingAction.ActionStep = actionStep.ActionStep;
+                                            existingAction.Status = actionStep.Status;
+                                        }
+                                    }
+                            }
                         }
                     }
                 }
@@ -493,7 +568,9 @@ namespace HOTP.Controllers
                 ViewData[_formData] = formCollection[_formData];
             }
             string copyFrom = Request["copyFrom"].ToString();
-            string copyTo= Request["copyTo"].ToString();
+            string copyTo = Request["copyTo"].ToString();
+            int copyFromInt = Convert.ToInt16(copyFrom);
+            int copyToInt = Convert.ToInt16(copyTo);
 
             for (int g = 0; g < 5; g++)
             {
@@ -505,16 +582,39 @@ namespace HOTP.Controllers
                 catch { }
                 if (empGoalID != null)
                 {
-                    int employeeGoalID = Convert.ToInt16(empGoalID);
-                    //var copyFrom = 
-                }
-            }
+                    try
+                    {
+                        int employeeGoalID = Convert.ToInt16(empGoalID);
+                        tblHOTP_Plan90 fromPlan = db.tblHOTP_Plan90.Where(p => p.EmployeeGoalID == employeeGoalID && p.Quarter == copyFromInt).First();
+                        tblHOTP_Plan90 toPlan = db.tblHOTP_Plan90.Where(p => p.EmployeeGoalID == employeeGoalID && p.Quarter == copyToInt).First();
+                        toPlan.Goal = fromPlan.Goal;
 
-                return RedirectToAction("Index", "Plan90");
+                        var toSteps = from step in db.tblHOTP_ActionSteps where step.PlanID == toPlan.PlanID select step;
+                        foreach (tblHOTP_ActionSteps step in toSteps)
+                        {
+                            db.tblHOTP_ActionSteps.Remove(step);
+                        }
+
+                        var fromSteps = from step in db.tblHOTP_ActionSteps where step.PlanID == fromPlan.PlanID select step;
+                        foreach (tblHOTP_ActionSteps step in fromSteps)
+                        {
+                            tblHOTP_ActionSteps newActionStep = new tblHOTP_ActionSteps();
+                            newActionStep.PlanID = toPlan.PlanID;
+                            newActionStep.ActionStep = step.ActionStep;
+                            newActionStep.Result = step.Result;
+                            newActionStep.Status = step.Status;
+                            db.tblHOTP_ActionSteps.Add(newActionStep);
+                        }
+                    }
+                    catch { }
+                }
+                db.SaveChanges();
+            }
+            return RedirectToAction("Index", "Plan90");
         }
 
 
-        
+
         protected override void Dispose(bool disposing)
         {
             if (disposing)
@@ -537,5 +637,30 @@ namespace HOTP.Controllers
 
 
     }
+
+    //public class BinaryContentResult : ActionResult
+    //{
+    //    private string ContentType;
+    //    private byte[] ContentBytes;
+
+    //    public BinaryContentResult(byte[] contentBytes, string contentType)
+    //    {
+    //        this.ContentBytes = contentBytes;
+    //        this.ContentType = contentType;
+    //    }
+
+    //    public override void ExecuteResult(ControllerContext context)
+    //    {
+    //        var response = context.HttpContext.Response;
+    //        response.Clear();
+    //        response.Cache.SetCacheability(HttpCacheability.NoCache);
+    //        response.ContentType = this.ContentType;
+
+    //        var stream = new MemoryStream(this.ContentBytes);
+    //        stream.WriteTo(response.OutputStream);
+    //        stream.Dispose();
+    //    }
+    //}
+
 
 }
